backend/requirements.txt
# LIFEASY V27+ PHASE 6 - Backend Dependencies
# Production-ready FastAPI Application with OTP, Chat, Notifications & Calling

# Web Framework
fastapi==0.135.1
uvicorn[standard]==0.41.0

# Database
sqlalchemy==2.0.48
psycopg2-binary==2.9.10  # PostgreSQL adapter

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# SMS/OTP Services
twilio==9.8.6
requests==2.32.3

# Push Notifications
firebase-admin==6.7.0

# Payment Gateway Support
httpx==0.28.1

# Configuration
pydantic==2.12.5
pydantic-settings==2.10.1
python-dotenv==1.1.0

# Production server management
gunicorn==23.0.0

# Development (optional)
# pytest==8.0.0
